requirements.txt
beautifulsoup4==4.9.3
bitarray==2.1.3
greenlet==1.1.0
install==1.3.4
msgpack==1.0.2
numpy==1.21.0
opencv-python==4.5.2.54
pdf2image==1.16.0
pdfreader==0.1.10
Pillow==8.2.0
pycryptodome==3.10.1
pynvim==0.4.3
pytesseract==0.3.7
python-dateutil==2.8.1
six==1.16.0
soupsieve==2.2.1
